requirements.txt
fastapi
uvicorn
groq
crawl4ai
beautifulsoup4
pydantic
python-dotenv
requests
jinja2